garvis/server/pyproject.toml
[project]
name = "garvis-server"
version = "0.1.0"
description = "Garvis XR Voice Assistant Server with Vision capabilities"
readme = "README.md"
requires-python = ">=3.11,<3.14"  # TensorFlow doesn't support 3.14 yet
dependencies = [
    # Core web framework and voice
    "aiohttp>=3.13.2",
    "anthropic>=0.75.0",
    "beautifulsoup4>=4.12.0",
    "deepgram-sdk>=5.3.0",
    "elevenlabs>=2.27.0",
    "fastapi>=0.127.1",
    "fastmcp>=2.14.1",
    "httpx>=0.28.1",
    "lxml>=5.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
    # Vision ML dependencies
    "ultralytics>=8.3.0",  # YOLOv8/v11
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "deepface>=0.0.89",  # Face detection (RetinaFace backend)
    "tf-keras>=2.15.0",  # Required by DeepFace
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
]
